src/geometry/lattice.py
"""
NovaTurbo — TPMS Lattice Structure Generator

Generates Triply Periodic Minimal Surface (TPMS) lattice structures
for engine components, similar to LEAP 71 / Noyron's approach.

Supported TPMS types:
  - Gyroid:    sin(x)cos(y) + sin(y)cos(z) + sin(z)cos(x) = t
  - Schwarz-P: cos(x) + cos(y) + cos(z) = t
  - Diamond:   sin(x)sin(y)sin(z) + sin(x)cos(y)cos(z) +
               cos(x)sin(y)cos(z) + cos(x)cos(y)sin(z) = t

The lattice is generated in cylindrical annular regions (matching engine
component geometries) and exported as STL meshes via marching cubes.
"""

import numpy as np
from dataclasses import dataclass, field
from typing import Tuple, Optional, List
from enum import Enum


class TPMSType(Enum):
    GYROID = "gyroid"
    SCHWARZ_P = "schwarz_p"
    DIAMOND = "diamond"


@dataclass
class LatticeParams:
    """Parameters for TPMS lattice generation."""
    tpms_type: TPMSType = TPMSType.GYROID

    # Cell size controls the lattice density
    cell_size_mm: float = 6.0       # Size of one unit cell in mm

    # Wall thickness of lattice struts
    wall_thickness: float = 0.3     # Iso-level threshold (0.1-0.5, higher = thicker)

    # Resolution (voxels per unit cell)
    resolution: int = 20            # Higher = smoother but slower

    # Density grading (varies lattice density along radius)
    grade_radially: bool = True     # Denser near hot inner wall
    inner_density: float = 0.4      # Thicker near inner wall (hotter)
    outer_density: float = 0.2      # Thinner near outer wall (cooler)

    # Density grading along axis
    grade_axially: bool = True
    upstream_density: float = 0.25   # Near compressor exit
    midstream_density: float = 0.4   # Primary combustion zone (thickest)
    downstream_density: float = 0.3  # Near turbine inlet


@dataclass
class AnnularRegion:
    """Defines an annular cylindrical region for lattice infill."""
    z_start_mm: float = 0.0
    z_end_mm: float = 80.0
    r_inner_mm: float = 30.0       # Inner wall radius
    r_outer_mm: float = 55.0       # Outer wall radius

    # Wall thicknesses to preserve (solid, no lattice)
    inner_wall_mm: float = 1.0     # Solid inner liner
    outer_wall_mm: float = 1.2     # Solid outer liner


def evaluate_tpms(X: np.ndarray, Y: np.ndarray, Z: np.ndarray,
                  tpms_type: TPMSType) -> np.ndarray:
    """Evaluate TPMS implicit function on a 3D grid."""
    if tpms_type == TPMSType.GYROID:
        return np.sin(X) * np.cos(Y) + np.sin(Y) * np.cos(Z) + np.sin(Z) * np.cos(X)
    elif tpms_type == TPMSType.SCHWARZ_P:
        return np.cos(X) + np.cos(Y) + np.cos(Z)
    elif tpms_type == TPMSType.DIAMOND:
        return (np.sin(X) * np.sin(Y) * np.sin(Z) +
                np.sin(X) * np.cos(Y) * np.cos(Z) +
                np.cos(X) * np.sin(Y) * np.cos(Z) +
                np.cos(X) * np.cos(Y) * np.sin(Z))
    else:
        raise ValueError(f"Unknown TPMS type: {tpms_type}")


def generate_annular_lattice(region: AnnularRegion,
                              params: LatticeParams) -> Tuple[np.ndarray, np.ndarray]:
    """
    Generate TPMS lattice that conforms to a cylindrical annular region.

    Strategy (slab-warp):
      1. Create a flat slab in (circumferential, radial, axial) space
      2. Evaluate TPMS on the regular slab grid → clean marching cubes
      3. Warp the slab into a cylinder so the lattice follows curvature
      4. Stitch the seam at θ=0/2π by merging nearby vertices

    This produces a lattice that naturally adapts to both inner and
    outer cylindrical walls, matching Noyron's conformal approach.
    """
    from skimage import measure
    import trimesh as _trimesh

    r_fill_inner = region.r_inner_mm + region.inner_wall_mm
    r_fill_outer = region.r_outer_mm - region.outer_wall_mm
    z_len = region.z_end_mm - region.z_start_mm

    if r_fill_outer <= r_fill_inner:
        raise ValueError("Walls too thick — no space for lattice infill")

    r_avg = (r_fill_inner + r_fill_outer) / 2.0
    radial_thickness = r_fill_outer - r_fill_inner
    circumference = 2 * np.pi * r_avg

    # Slab dimensions in mm
    # u = circumferential [0, circumference + overlap]
    # v = radial [0, radial_thickness]
    # w = axial [0, z_len]
    overlap = params.cell_size_mm * 1.2  # overlap past full circle for seam stitching

    voxels_per_mm = params.resolution / params.cell_size_mm
    nu = max(int((circumference + overlap) * voxels_per_mm), 60)
    nv = max(int(radial_thickness * voxels_per_mm), 12)
    nw = max(int(z_len * voxels_per_mm), 30)

    # Cap to avoid memory explosion
    nu = min(nu, 300)
    nv = min(nv, 60)
    nw = min(nw, 150)

    u = np.linspace(0, circumference + overlap, nu)
    v = np.linspace(0, radial_thickness, nv)
    w = np.linspace(0, z_len, nw)
    U, V, W = np.meshgrid(u, v, w, indexing='ij')

    # Evaluate TPMS in slab space (perfectly Cartesian → clean topology)
    scale = 2 * np.pi / params.cell_size_mm
    field = evaluate_tpms(U * scale, V * scale, W * scale, params.tpms_type)

    spacing = (
        (circumference + overlap) / (nu - 1),
        radial_thickness / (nv - 1),
        z_len / (nw - 1),
    )

    try:
        verts, faces, normals, values = measure.marching_cubes(
            field, level=0.0, spacing=spacing
        )
    except (ValueError, RuntimeError):
        return np.array([]).reshape(0, 3), np.array([]).reshape(0, 3)

    # verts are in (u, v, w) slab space — warp to cylinder
    u_v = verts[:, 0]   # circumferential position (mm along arc)
    v_v = verts[:, 1]   # radial offset from inner fill wall
    w_v = verts[:, 2]   # axial offset

    theta = u_v / r_avg  # convert arc length to angle
    r = r_fill_inner + v_v
    z = region.z_start_mm + w_v

    # Cylindrical to Cartesian
    x = r * np.cos(theta)
    y = r * np.sin(theta)

    verts_cyl = np.column_stack([x, y, z])

    # Clip faces in the overlap region (keep only θ < 2π)
    face_centers_u = verts[:, 0][faces].mean(axis=1)
    keep = face_centers_u <= circumference
    clipped_faces = faces[keep]

    if len(clipped_faces) == 0:
        return np.array([]).reshape(0, 3), np.array([]).reshape(0, 3)

    mesh = _trimesh.Trimesh(vertices=verts_cyl, faces=clipped_faces)
    mesh.remove_unreferenced_vertices()

    # Stitch the seam: merge vertices that overlap at θ≈0 and θ≈2π
    mesh.merge_vertices(merge_tex=True, merge_norm=True)
    mesh.fix_normals()

    return np.array(mesh.vertices), np.array(mesh.faces)


def generate_combustor_lattice(combustor_params=None,
                                z_offset: float = 0.0,
                                lattice_params: LatticeParams = None) -> Tuple[np.ndarray, np.ndarray]:
    """
    Generate TPMS lattice for the combustor liner walls.
    This replaces the solid liner walls with a lattice-infilled structure
    for superior heat transfer and weight reduction.
    """
    if combustor_params is None:
        from .combustor import CombustorParams
        combustor_params = CombustorParams()

    if lattice_params is None:
        lattice_params = LatticeParams()

    # Lattice fills between inner liner and outer casing wall
    # Must stay strictly inside the combustor shell (r=28.8 to 57.6)
    outer_region = AnnularRegion(
        z_start_mm=z_offset,
        z_end_mm=z_offset + combustor_params.length_mm,
        r_inner_mm=combustor_params.inner_diameter_mm / 2.0 + combustor_params.liner_thickness_mm,
        r_outer_mm=combustor_params.outer_diameter_mm / 2.0 - combustor_params.casing_thickness_mm,
        inner_wall_mm=0.5,
        outer_wall_mm=0.5,
    )

    all_verts = []
    all_faces = []
    face_offset = 0

    for region in [outer_region]:
        try:
            verts, faces = generate_annular_lattice(region, lattice_params)
            if len(verts) > 0:
                all_verts.append(verts)
                all_faces.append(faces + face_offset)
                face_offset += len(verts)
        except ValueError as e:
            print(f"  Warning: Skipping region — {e}")

    if not all_verts:
        return np.array([]).reshape(0, 3), np.array([]).reshape(0, 3)

    return np.vstack(all_verts), np.vstack(all_faces)


def generate_nozzle_lattice(nozzle_params=None,
                             z_offset: float = 0.0,
                             lattice_params: LatticeParams = None) -> Tuple[np.ndarray, np.ndarray]:
    """
    Generate TPMS lattice for nozzle walls.
    Lighter lattice since thermal loads are lower here.
    """
    if nozzle_params is None:
        from .nozzle import NozzleParams
        nozzle_params = NozzleParams()

    if lattice_params is None:
        lattice_params = LatticeParams(
            cell_size_mm=5.0,
            wall_thickness=0.25,
            grade_radially=True,
            inner_density=0.3,
            outer_density=0.15,
            grade_axially=False,
        )

    # Nozzle is a converging cone — lattice fills the wall thickness
    r_inlet = nozzle_params.inlet_diameter_mm / 2.0
    r_exit = nozzle_params.exit_diameter_mm / 2.0
    wall_t = nozzle_params.wall_thickness_mm if hasattr(nozzle_params, 'wall_thickness_mm') else 2.0
    # Use average radius, lattice fills inside the wall
    avg_r = (r_inlet + r_exit) / 2.0

    region = AnnularRegion(
        z_start_mm=z_offset,
        z_end_mm=z_offset + nozzle_params.length_mm,
        r_inner_mm=avg_r - wall_t * 2,
        r_outer_mm=avg_r + wall_t * 2,
        inner_wall_mm=0.3,
        outer_wall_mm=0.3,
    )

    try:
        return generate_annular_lattice(region, lattice_params)
    except ValueError:
        return np.array([]).reshape(0, 3), np.array([]).reshape(0, 3)


def get_lattice_stats(verts: np.ndarray, faces: np.ndarray,
                       region: AnnularRegion = None) -> dict:
    """Compute statistics for the generated lattice."""
    if len(verts) == 0:
        return {'vertices': 0, 'faces': 0, 'volume_ratio': 0}

    stats = {
        'vertices': len(verts),
        'faces': len(faces),
        'bounds_min': verts.min(axis=0).tolist(),
        'bounds_max': verts.max(axis=0).tolist(),
    }

    if region:
        total_vol = (np.pi * (region.r_outer_mm**2 - region.r_inner_mm**2) *
                     (region.z_end_mm - region.z_start_mm))
        # Rough volume ratio from lattice fill fraction
        stats['total_annular_volume_mm3'] = total_vol
        stats['estimated_fill_fraction'] = 0.3  # Typical for TPMS

    return stats


if __name__ == "__main__":
    print("=== NovaTurbo TPMS Lattice Generator ===\n")

    params = LatticeParams(
        tpms_type=TPMSType.GYROID,
        cell_size_mm=6.0,
        wall_thickness=0.3,
        resolution=20,
    )

    print(f"  Type: {params.tpms_type.value}")
    print(f"  Cell size: {params.cell_size_mm} mm")
    print(f"  Resolution: {params.resolution} voxels/cell")
    print(f"  Grading: radial={'yes' if params.grade_radially else 'no'}, "
          f"axial={'yes' if params.grade_axially else 'no'}")

    print("\n  Generating combustor lattice...")
    verts, faces = generate_combustor_lattice(lattice_params=params)
    print(f"  ✓ {len(verts):,} vertices, {len(faces):,} faces")

    if len(verts) > 0:
        stats = get_lattice_stats(verts, faces)
        print(f"  Bounds: {stats['bounds_min']} → {stats['bounds_max']}")
